total_paid.py handler is singled out.
