+++ bistro/orders/views.py
@@ -1,11 +1,5 @@
-from datetime import datetime
-from decimal import Decimal
-
-from django.db.models import F
-from django.db.models import Sum
 from django.shortcuts import get_object_or_404
 from django.shortcuts import render
-from django.utils import timezone
 
 from .forms import DeleteOrderForm
 from .forms import OrderForm
@@ -15,6 +9,7 @@ from .forms import UpdateStatusForm
 from .handlers.add_order import handle_add_order
 from .handlers.delete_order import handle_delete_order
 from .handlers.search_order import handle_search_order
+from .handlers.total_paid import handle_total_paid_orders
 from .handlers.update_status import handle_update_status
 from .models import Order
 
@@ -45,29 +40,6 @@ def order_management_view(request):
         "modals": get_modals(),
     }
     return render(request, "orders/order_management.html", context)
-
-
-def handle_total_paid_orders(request):
-    form = TotalPaidOrdersForm(request.POST)
-    if form.is_valid():
-        date = form.cleaned_data["date"]
-
-        start_datetime = timezone.make_aware(
-            datetime.combine(date, datetime.min.time()),
-        )
-        end_datetime = timezone.make_aware(datetime.combine(date, datetime.max.time()))
-
-        total_paid = Order.objects.filter(
-            status="paid",
-            created_at__range=[start_datetime, end_datetime],
-        ).annotate(
-            total_order_price=Sum(
-                F("order_items__menu_item__price") * F("order_items__quantity"),
-            ),
-        ).aggregate(total=Sum("total_order_price"))["total"] or Decimal("0.00")
-
-        return total_paid  # noqa: RET504
-    return None
 
 
 def get_modals():
